Setup & add notes for first lecture
Sets up the baseline project: Adds `latex-packages` as a submodule and
puts the basic source file infrastructure (`[...].tex`, `latexmkrc`,
`.gitignore`) under the subdirectory `src`. Moreover, this commit also
contains substantial work on the notes, including a preliminary
transcript of the first lecture (among other things, e.g. a disclaimer
and an explanation of document structure, a bibliography file and so
on).