Add rough work on lectures 3, 4, 5 & bump `latex-packages`
I apologize for being very sloppy with actually committing any work.
This commit adds a veritable mountain of text and some changes to
already existing material, mainly to the only spectral sequence
"illustration" that we have right now.

We also bump `latex-packages` by a few commits which in itself fixes
some bugs, in particular with the color of some `mdframed` lines being
context-dependent.