Add exercises & (most) solutions
This commit dumps all of the exercises in and adds solutions to
everything we got correct, leaving the few remaining for later. We also
make sure all hboxes are fine and do some minor work here and there
towards integrating the exercises with the main text body, but there is
still a lot to do on that front.