Add exercises up to 9 & recheck introduction
Starting on the process of rereading, checking, and improving everything
I wrote. This commit covers the "pre-introduction" / disclaimer and the
Informal Introduction. To be able to cite some structural theorem we
proved later on and in the exercises, we add all the remaining corrected
exercises up to this point and also attach labels to the relevant
things.

We also add a new citation, namely the lecture notes by Davis & Kirk,
which are both wonderful and the first reference to the
Alexander-Whitney map I was able to find (although I didn't bother to
check either tom Dieck or Bredon).