Add TikZ externalization & small improvements
Finally got around to enabling externalization for tikzpictures since
compilation has gotten excruciatingly slow. I feel retroactively
justified in putting it off for so long, too: Getting this to work took
a lot of poking things with a stick in the dark and a bunch of "passive
help" from the internet (see, say, the changes to `latexmkrc` made).

We also make very minor improvements by rewording here and there in
`spectral_sequences.tex`.