`spectral_sequences`: Rework up to "Construction"
Reworks and corrects, and most importantly fills in all missing
illustrations for the everything in the spectral sequences section up to
but not including the construction.

Of course, this involves adding some things, mostly TikZ definitions, to
the preamble as well, and finally we also make use of some changes to
`latex-packages`.