Rework "Rational homotopy groups" section
A few corrections, a few pictures here and there, nothing out of the
usual. We also expand our TikZ definitions a little to finally
streamline setting differentials, and sprinkle a few labels on whatever
result or exercise we now cite.

At least one open factual question and one optional picture for which I
haven't quite decided yet whether it is actually worth the additional
work remain, so section this will have to be revisited later down the
line.